requirements.txt
parquet
pandas
pyarrow
transformers
bitsandbytes
accelerate
pillow
diffusers[torch]
sentencepiece
beautifulsoup4
peft
--extra-index-url https://download.pytorch.org/whl/cu121
torch